requirements.txt
# Core
numpy>=1.24
opencv-python>=4.8
pyyaml>=6.0
tqdm>=4.66

# Training / inference
torch>=2.1
torchvision>=0.16

# Detectron2 (install separately if pip fails; see README)
detectron2>=0.6

# For docs
python-docx>=1.1

pycocotools>=2.0.7
